fix: Error enum is assembled in "Error file"
